god mode ca fatigue mais toujours al, fork need fix

+++ VM/src/operations/zjmp_fct.c
@@ -27,7 +27,7 @@ int			zjmp_fct(t_war *war, t_chariot *chariot)
 	{
 		war->back_pc = 0;
 		 if (war->verbose[2] == 1)
-		 	printf("P %4d | zjmp %d FAILED\n", (chariot->index + 1), (short)war->param[0]);
+		 	printf("P %4d | zjmp %d FAILED\n", (chariot->index + 1), war->param[0]);
  	}
  	return (SUCCESS);
 }
